Warning removal.
git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@1920 7894878c-1315-0410-8ee3-d5d059ff63e0

+++ prog/sensord/rrd.c
@@ -306,7 +306,7 @@ rrdCGI_LINE
 (void *_data, const char *rawLabel, const char *label, const FeatureDescriptor *feature) {
   struct gr *data = (struct gr *) _data;
   if (!feature || (feature->rrd && (feature->type == data->type)))
-    printf ("\n\tLINE2:%s#%.6x:\"%s\"", rawLabel, random () & 0xffffff, label);
+    printf ("\n\tLINE2:%s#%.6x:\"%s\"", rawLabel, (int) random () & 0xffffff, label);
   return 0;
 }
